% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detail.R
\name{get_detail}
\alias{get_detail}
\title{GET Details for a CAS Registry Number}
\usage{
get_detail(cas_rn, json = FALSE)
}
\arguments{
\item{cas_rn}{(Character) CAS Registry Number with dashes, e.g., "50-78-2".}

\item{json}{(Logical) Should the result be returned as JSON? Defaults to
\code{FALSE}.}
}
\value{
Returns a list.
}
\description{
This function performs a query to retrieve details for a CAS
  Registry Number.
}
\details{
The function performs a sanity check on the provided CAS Registry
  Number and then performs a query. If successful, a list with the available
  details will be returned.
}
\examples{
\dontrun{
cas_rn <- "50-78-2"
get_detail(cas_rn)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
