% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_export.R
\name{get_export}
\alias{get_export}
\title{GET the SDF Export for a CAS Registry Number}
\usage{
get_export(cas_rn, attachment = FALSE)
}
\arguments{
\item{cas_rn}{(Character) CAS Registry Number in "dash" format, e.g.,
"50-78-2".}

\item{attachment}{(Logical) Should the result be returned as attachment?
Currently only \code(FALSE) is supported.}
}
\value{
Returns a character string.
}
\description{
This function performs a query to retrieve a SDF file for a CAS
  Registry Number.
}
\details{
The function performs a sanity check on the provided CAS Registry
  Number and then performs a query. If successful, a character string with
  the available SDF file will be returned.
}
\examples{
\dontrun{
cas_rn <- "50-78-2"
get_export(cas_rn)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
