% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_search.R
\name{get_search}
\alias{get_search}
\title{GET Search Results for a CAS Registry Number and Other Identifiers}
\usage{
get_search(q, offset = NULL, size = NULL)
}
\arguments{
\item{q}{(Character) The (partial) search term. The following search terms
are supported: CAS Registry Number (with or without dashes), SMILES
(canonical or isomeric), InChI (with or without the "InChI=" prefix),
InChIKey, and name. Searching by name allows use of a trailing wildcard
(e.g., car*). All searches are case-insensitive.}

\item{offset}{(NULL) Currently not implemented.}

\item{size}{(NULL) Currently not implemented.}
}
\value{
Returns a list.
}
\description{
This function performs a query to search for substances based
  on a CAS Registry Number or other identifiers.
}
\details{
The function performs a sanity check on the provided search term
  and then performs a query. If successful, a list with the available
  information will be returned.
}
\note{
When searching for an InChIKey, the prefix \code{"InChIKey="} needs to
  be added.
}
\examples{
\dontrun{
q <- "Aspi*"
get_search(q)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
